ENH: High density display scaling.
From Qt 5.6 pixels are logical, this means that under some circumstances
they will be multiplied by a given factor before they are rendered.

This feature can be used to increase font size and icon sizes on high
density displays (like medical displays).

However, VTK does not support it well, so an special care has to be
taken because this multiplication factor does not apply on OpenGL
widgets.

For this reason, mouse events sent to VTK are in logical pixels, so they
must be multiplied.

IMPORTANT:

As soon as VTK supports pixel scaling, the modifications on QViewer
should be removed.

This new feature is still experimental.