ENH: new profiling tools
New classes have been implemented in order to profile the time it takes
to run some parts of the code.

Profilers:

A central point where profilers can be added and retreived anywhere in
the code. It's somewhat like a singleton.

Profiler:

Where you can get tasks. This class controls the amount of repeated task
names and stores all of them.

Task:

Counts the time elapsed from it's creation until finish is called.

Printer:

Abstract class assigned to a profiler that receives the events that
happen on the Profiling (task started, task finished ...).

TextPrinter:

A real implementation of Printer that outputs something readable in text format.